refactor[#14] : package 이동 및 test method 명 변경
